Rebase and move cert checker to new ct package.
