+++ examples/ct/cert_checker_test.go
@@ -1,4 +1,4 @@
-package ct_fe
+package ct
 
 import (
 	"encoding/pem"
